+++ src/screens/Messages/ConversationSettings/constants.ts
@@ -1,0 +1,2 @@
+export const ROW_SPACING = 10
+export const MEMBER_LIMIT = 50
