+++ src/screens/Messages/ConversationSettings/index.tsx
@@ -1,0 +1,552 @@
+import {useState} from 'react'
+import {View} from 'react-native'
+import {Trans, useLingui} from '@lingui/react/macro'
+import {StackActions, useNavigation} from '@react-navigation/native'
+
+import {useBottomBarOffset} from '#/lib/hooks/useBottomBarOffset'
+import {useInitialNumToRender} from '#/lib/hooks/useInitialNumToRender'
+import {
+  type CommonNavigatorParams,
+  type NativeStackScreenProps,
+  type NavigationProp,
+} from '#/lib/routes/types'
+import {logger} from '#/logger'
+import {ConvoProvider, useConvo} from '#/state/messages/convo'
+import {ConvoStatus} from '#/state/messages/convo/types'
+import {useEditGroupChatName} from '#/state/queries/messages/edit-group-chat-name'
+import {useLeaveConvo} from '#/state/queries/messages/leave-conversation'
+import {useListJoinRequestsQuery} from '#/state/queries/messages/list-join-requests'
+import {useMuteConvo} from '#/state/queries/messages/mute-conversation'
+import {useSession} from '#/state/session'
+import {List} from '#/view/com/util/List'
+import {atoms as a, useBreakpoints, useTheme} from '#/alf'
+import {AvatarBubbles} from '#/components/AvatarBubbles'
+import {Button, type ButtonColor, ButtonIcon} from '#/components/Button'
+import * as Dialog from '#/components/Dialog'
+import {type ConvoWithDetails, parseConvoView} from '#/components/dms/util'
+import {Error} from '#/components/Error'
+import {ArrowBoxLeft_Stroke2_Corner0_Rounded as ArrowBoxLeftIcon} from '#/components/icons/ArrowBoxLeft'
+import {
+  Bell2_Stroke2_Corner0_Rounded as BellIcon,
+  Bell2Off_Stroke2_Corner0_Rounded as BellOffIcon,
+} from '#/components/icons/Bell2'
+import {ChainLink_Stroke2_Corner0_Rounded as ChainLinkIcon} from '#/components/icons/ChainLink'
+import {type Props as SVGIconProps} from '#/components/icons/common'
+import {DotGrid3x1_Stroke2_Corner0_Rounded as EllipsisIcon} from '#/components/icons/DotGrid'
+import {EditBig_Stroke2_Corner2_Rounded as EditIcon} from '#/components/icons/EditBig'
+import {Flag_Stroke2_Corner0_Rounded as FlagIcon} from '#/components/icons/Flag'
+import {Lock_Stroke2_Corner0_Rounded as LockIcon} from '#/components/icons/Lock'
+import * as Layout from '#/components/Layout'
+import * as Prompt from '#/components/Prompt'
+import * as Toast from '#/components/Toast'
+import {Text} from '#/components/Typography'
+import {IS_NATIVE} from '#/env'
+import type * as bsky from '#/types/bsky'
+import {InviteLinkDialog} from '../components/InviteLinkDialog'
+import {AddMembersLink} from './AddMembersLink'
+import {ROW_SPACING} from './constants'
+import {Member} from './Member'
+import {MembersAndRequests} from './MembersAndRequests'
+import {EditNamePrompt, LeaveChatPrompt, LockChatPrompt} from './prompts'
+
+const dateFormatter = new Intl.DateTimeFormat(undefined, {
+  month: 'long',
+  day: 'numeric',
+  year: 'numeric',
+})
+
+type Item =
+  | {
+      type: 'MEMBERS_AND_REQUESTS'
+    }
+  | {
+      type: 'ADD_MEMBERS_LINK'
+    }
+  | {
+      type: 'CHAT_MEMBER'
+      profile: bsky.profile.AnyProfileView
+      status: 'owner' | 'standard' | 'invited'
+    }
+
+type Props = NativeStackScreenProps<
+  CommonNavigatorParams,
+  'MessagesConversationSettings'
+>
+
+export function MessagesConversationSettingsScreen({route}: Props) {
+  const {gtTablet} = useBreakpoints()
+
+  const convoId = route.params.conversation
+
+  return (
+    <Layout.Screen>
+      <Layout.Header.Outer>
+        <Layout.Header.BackButton />
+        <Layout.Header.Content align={gtTablet ? 'left' : 'platform'}>
+          <Layout.Header.TitleText>
+            <Trans>Group chat settings</Trans>
+          </Layout.Header.TitleText>
+        </Layout.Header.Content>
+        <Layout.Header.Slot />
+      </Layout.Header.Outer>
+      <ConvoProvider key={convoId} convoId={convoId}>
+        <SettingsInner convoId={convoId} />
+      </ConvoProvider>
+    </Layout.Screen>
+  )
+}
+
+function keyExtractor(item: Item) {
+  return item.type === 'CHAT_MEMBER' ? item.profile.did : item.type
+}
+
+function SettingsInner({convoId}: {convoId: string}) {
+  const {t: l} = useLingui()
+
+  const initialNumToRender = useInitialNumToRender({minItemHeight: 68})
+  const bottomBarOffset = useBottomBarOffset()
+
+  const convoState = useConvo()
+  const {currentAccount} = useSession()
+
+  const convo = convoState.convo
+    ? parseConvoView(convoState.convo, currentAccount?.did)
+    : null
+  const primaryMember = convo?.primaryMember
+  const isOwner = !!primaryMember && primaryMember.did === currentAccount?.did
+
+  const data: bsky.profile.AnyProfileView[] = convo?.members ?? []
+  // TODO Need this data in order to populate this array. -dsb
+  const invites: string[] = []
+
+  const {data: joinRequestsData, hasNextPage: hasMoreRequests} =
+    useListJoinRequestsQuery({
+      convoId,
+      enabled: isOwner,
+    })
+  const requestCount =
+    joinRequestsData?.pages.reduce(
+      (sum, page) => sum + page.requests.length,
+      0,
+    ) ?? 0
+
+  const items: Item[] = [
+    {
+      type: 'MEMBERS_AND_REQUESTS',
+    },
+    ...(isOwner ? [{type: 'ADD_MEMBERS_LINK'} as const] : []),
+    ...[...data]
+      .sort((a, b) => {
+        const aIsOwner = a.did === primaryMember?.did
+        const bIsOwner = b.did === primaryMember?.did
+        const aIsSelf = a.did === currentAccount?.did
+        const bIsSelf = b.did === currentAccount?.did
+        if (aIsOwner !== bIsOwner) return aIsOwner ? -1 : 1
+        if (aIsSelf !== bIsSelf) return aIsSelf ? -1 : 1
+        return 0
+      })
+      .map(
+        (profile): Item => ({
+          type: 'CHAT_MEMBER',
+          profile,
+          status:
+            primaryMember?.did === profile.did
+              ? 'owner'
+              : invites.includes(profile.did)
+                ? 'invited'
+                : 'standard',
+        }),
+      ),
+  ]
+
+  function renderItem({item}: {item: Item}) {
+    switch (item.type) {
+      case 'MEMBERS_AND_REQUESTS':
+        return (
+          <MembersAndRequests
+            memberCount={data.length}
+            requestCount={requestCount}
+            hasMoreRequests={!!hasMoreRequests}
+            isOwner={isOwner}
+          />
+        )
+      case 'ADD_MEMBERS_LINK':
+        return convo ? (
+          <AddMembersLink
+            convo={convo}
+            members={data.map(profile => profile.did)}
+          />
+        ) : null
+      case 'CHAT_MEMBER':
+        return convo ? (
+          <Member
+            convo={convo}
+            profile={item.profile}
+            status={item.status}
+            isOwner={isOwner}
+          />
+        ) : null
+      default:
+        return null
+    }
+  }
+
+  if (convoState.status === ConvoStatus.Error) {
+    return (
+      <Error
+        title={l`Something went wrong`}
+        message={l`We couldn’t load this conversation’s settings`}
+        onRetry={() => convoState.error.retry()}
+        sideBorders={false}
+      />
+    )
+  }
+
+  return (
+    <List
+      data={items}
+      contentContainerStyle={{paddingBottom: bottomBarOffset + ROW_SPACING}}
+      desktopFixedHeight
+      initialNumToRender={initialNumToRender}
+      keyExtractor={keyExtractor}
+      ListHeaderComponent={
+        convo?.kind === 'group' ? (
+          <SettingsHeader convo={convo} isOwner={isOwner} />
+        ) : (
+          <SettingsHeaderPlaceholder />
+        )
+      }
+      renderItem={renderItem}
+      sideBorders={false}
+      windowSize={11}
+      // TODO Paginate on relatedProfiles. -dsb
+      onEndReached={() => {}}
+      onEndReachedThreshold={IS_NATIVE ? 1.5 : 0}
+    />
+  )
+}
+
+function SettingsHeader({
+  convo,
+  isOwner,
+}: {
+  convo: Extract<ConvoWithDetails, {kind: 'group'}>
+  isOwner: boolean
+}) {
+  const t = useTheme()
+  const {t: l} = useLingui()
+
+  const navigation = useNavigation<NavigationProp>()
+
+  const groupName = convo.details.name
+  const [newGroupName, setNewGroupName] = useState(groupName)
+
+  const [isLocked, setIsLocked] = useState(false)
+
+  // TODO Enable this once the feature is working end-to-end. -dsb
+  // const {joinLink} = convo.details
+  const isJoinLinkEnabled = false
+  // const isJoinLinkEnabled =
+  //   isOwner || (!isOwner && joinLink?.enabledStatus === 'enabled')
+
+  // TODO Enable this once the feature is working end-to-end. -dsb
+  const isReportLinkEnabled = false
+
+  const {mutate: editGroupName} = useEditGroupChatName(convo.view.id, {
+    onError: e => {
+      setNewGroupName(groupName)
+      logger.error('Failed to edit group chat name', {message: e})
+      Toast.show(l`Failed to edit group chat name`, {
+        type: 'error',
+      })
+    },
+  })
+
+  const {mutate: muteConvo} = useMuteConvo(convo.view.id, {
+    onSuccess: data => {
+      if (data.convo.muted) {
+        Toast.show(l({message: 'Group chat muted', context: 'toast'}))
+      } else {
+        Toast.show(l({message: 'Group chat unmuted', context: 'toast'}))
+      }
+    },
+    onError: e => {
+      logger.error('Failed to mute group chat', {message: e})
+      Toast.show(l`Failed to mute group chat`, {
+        type: 'error',
+      })
+    },
+  })
+
+  const {mutate: leaveConvo} = useLeaveConvo(convo.view.id, {
+    onSuccess: () => {
+      // Settings > Chat > Chat list
+      navigation.dispatch(StackActions.pop(2))
+    },
+    onError: e => {
+      logger.error('Failed to leave group chat', {message: e})
+      Toast.show(l({message: 'Failed to leave group chat', context: 'toast'}), {
+        type: 'error',
+      })
+    },
+  })
+
+  const inviteLinkDialog = Dialog.useDialogControl()
+  const editNamePrompt = Prompt.usePromptControl()
+  const lockChatPrompt = Prompt.usePromptControl()
+  const leaveChatPrompt = Prompt.usePromptControl()
+
+  const handleToggleMute = () => {
+    muteConvo({mute: !convo.view.muted})
+  }
+
+  // TODO Need to implement this when the backend is ready. -dsb
+  const handleReportChat = () => {}
+
+  const handlePromptName = () => {
+    setNewGroupName(groupName)
+    editNamePrompt.open()
+  }
+
+  const handleEditName = () => {
+    editGroupName({name: newGroupName})
+  }
+
+  const handleConfirmLock = () => {
+    setIsLocked(true)
+  }
+
+  const handleUnlock = () => {
+    setIsLocked(false)
+  }
+
+  // TODO The creation date doesn't exist yet. -dsb
+  const showCreatedAt = false
+  const createdAt = new Date()
+
+  // TODO Need to implement this. -dsb
+  const canLockGroupChat = false
+  // const canLockGroupChat = isOwner
+
+  return (
+    <>
+      <View
+        style={[a.px_xl, a.py_4xl, a.border_b, t.atoms.border_contrast_low]}>
+        <View style={[a.align_center, a.justify_center]}>
+          <AvatarBubbles profiles={convo.members} />
+        </View>
+        <Text
+          style={[
+            a.text_2xl,
+            a.font_bold,
+            a.text_center,
+            a.pt_lg,
+            t.atoms.text,
+          ]}>
+          {groupName}
+        </Text>
+        {showCreatedAt ? (
+          <Text
+            style={[
+              a.text_sm,
+              a.text_center,
+              a.pt_xs,
+              a.px_xl,
+              t.atoms.text_contrast_high,
+            ]}>
+            <Trans>Created {dateFormatter.format(createdAt)}</Trans>
+          </Text>
+        ) : null}
+        <View
+          style={[
+            a.flex_row,
+            a.align_center,
+            a.justify_center,
+            a.gap_2xl,
+            a.pt_2xl,
+          ]}>
+          <SettingsButton
+            color={convo.view.muted ? 'negative_subtle' : 'secondary'}
+            icon={convo.view.muted ? BellOffIcon : BellIcon}
+            label={
+              convo.view.muted
+                ? l`Unmute this group chat`
+                : l`Mute this group chat`
+            }
+            text={convo.view.muted ? l`Muted` : l`Mute`}
+            onPress={handleToggleMute}
+          />
+          {isOwner ? (
+            <SettingsButton
+              icon={EditIcon}
+              label={l`Edit this group chat’s name`}
+              text={l`Edit name`}
+              onPress={handlePromptName}
+            />
+          ) : null}
+          {isJoinLinkEnabled ? (
+            <SettingsButton
+              icon={ChainLinkIcon}
+              label={
+                isOwner
+                  ? l`Create or modify an invite link for this group chat`
+                  : l`View the invite link for this group chat`
+              }
+              text={l`Invite link`}
+              onPress={inviteLinkDialog.open}
+            />
+          ) : null}
+          {canLockGroupChat ? (
+            <SettingsButton
+              color={isLocked ? 'negative_subtle' : 'secondary'}
+              icon={LockIcon}
+              label={
+                isLocked ? l`Unlock this group chat` : l`Lock this group chat`
+              }
+              text={isLocked ? l`Locked` : l`Lock`}
+              onPress={isLocked ? handleUnlock : lockChatPrompt.open}
+            />
+          ) : null}
+          {isOwner ? null : isReportLinkEnabled ? (
+            <SettingsButton
+              color="secondary"
+              icon={FlagIcon}
+              label={l`Report this group chat`}
+              text={l`Report`}
+              onPress={handleReportChat}
+            />
+          ) : null}
+          {isOwner ? null : (
+            <SettingsButton
+              color="secondary"
+              icon={ArrowBoxLeftIcon}
+              label={l`Leave this group chat`}
+              text={l`Leave`}
+              onPress={leaveChatPrompt.open}
+            />
+          )}
+        </View>
+      </View>
+      <EditNamePrompt
+        control={editNamePrompt}
+        value={newGroupName}
+        onChangeText={setNewGroupName}
+        onConfirm={handleEditName}
+      />
+      <InviteLinkDialog
+        convo={convo}
+        control={inviteLinkDialog}
+        isOwner={isOwner}
+      />
+      <LockChatPrompt control={lockChatPrompt} onConfirm={handleConfirmLock} />
+      <LeaveChatPrompt
+        control={leaveChatPrompt}
+        groupName={groupName}
+        onConfirm={leaveConvo}
+      />
+    </>
+  )
+}
+
+function SettingsHeaderPlaceholder() {
+  const t = useTheme()
+
+  return (
+    <View style={[a.px_xl, a.py_4xl, a.border_b, t.atoms.border_contrast_low]}>
+      <View style={[a.align_center, a.justify_center]}>
+        <AvatarBubbles profiles={[]} />
+      </View>
+      <Text
+        style={[a.text_2xl, a.font_bold, a.text_center, a.pt_lg, t.atoms.text]}>
+        …
+      </Text>
+      <Text
+        style={[
+          a.text_sm,
+          a.text_center,
+          a.pt_xs,
+          a.px_xl,
+          t.atoms.text_contrast_high,
+        ]}>
+        …
+      </Text>
+      <View
+        style={[
+          a.flex_row,
+          a.align_center,
+          a.justify_center,
+          a.gap_2xl,
+          a.pt_2xl,
+        ]}>
+        <SettingsButtonPlaceholder />
+        <SettingsButtonPlaceholder />
+        <SettingsButtonPlaceholder />
+        <SettingsButtonPlaceholder />
+      </View>
+    </View>
+  )
+}
+
+function SettingsButton({
+  color = 'secondary',
+  icon,
+  label,
+  text,
+  onPress,
+}: {
+  color?: ButtonColor
+  icon: React.ComponentType<SVGIconProps>
+  label: string
+  text: string
+  onPress: () => void
+}) {
+  const t = useTheme()
+
+  return (
+    <View style={[a.align_center]}>
+      <Button
+        color={color}
+        size="large"
+        shape="round"
+        label={label}
+        onPress={onPress}>
+        <ButtonIcon icon={icon} size="md" />
+      </Button>
+      <Text
+        numberOfLines={1}
+        style={[
+          a.text_2xs,
+          a.font_medium,
+          a.text_center,
+          a.pt_xs,
+          t.atoms.text,
+        ]}>
+        {text}
+      </Text>
+    </View>
+  )
+}
+
+function SettingsButtonPlaceholder() {
+  const t = useTheme()
+  const {t: l} = useLingui()
+
+  return (
+    <View>
+      <Button color="secondary" size="large" shape="round" label={l`Loading…`}>
+        <ButtonIcon icon={EllipsisIcon} size="md" />
+      </Button>
+      <Text
+        numberOfLines={1}
+        style={[
+          a.text_2xs,
+          a.font_medium,
+          a.text_center,
+          a.pt_xs,
+          t.atoms.text,
+        ]}>
+        …
+      </Text>
+    </View>
+  )
+}
